third_party/blink/web_tests/platform/win11-arm64/external/wpt/webxr/ar-module/xrSession_interactionMode.https-expected.txt

This is a testharness.js-based test.
[FAIL] Tests interactionMode for an VR_HMD_DEVICE - webgl
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] Tests interactionMode for an VR_HMD_DEVICE - webgl2
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] Tests interactionMode for an VR_SCREEN_DEVICE - webgl
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] Tests interactionMode for an VR_SCREEN_DEVICE - webgl2
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] Tests interactionMode for an AR_HMD_DEVICE - webgl
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] Tests interactionMode for an AR_HMD_DEVICE - webgl2
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] Tests interactionMode for an AR_SCREEN_DEVICE - webgl
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] Tests interactionMode for an AR_SCREEN_DEVICE - webgl2
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] Tests interactionMode for a INLINE_SCREEN_DEVICE - webgl
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] Tests interactionMode for a INLINE_SCREEN_DEVICE - webgl2
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
Harness: the test ran to completion.
